refactor: 리액트 컴포넌트와 스타일 컴포넌트를 분리
