+++ src/components/CategoryImage/CategoryImage.styles.ts
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const CategoryImageWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 64px;
+  height: 64px;
+  min-width: 64px;
+  min-height: 64px;
+
+  margin-right: 16px;
+
+  border-radius: 50%;
+  background: var(--lighten-color);
+
+  img {
+    width: 36px;
+    height: 36px;
+  }
+`;
