+++ src/components/Header/Header.styles.ts
@@ -1,0 +1,37 @@
+import styled from 'styled-components';
+
+import { textTitle } from '../../style/mixin';
+
+export const Header = styled.header`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 64px;
+
+  padding: 0 16px;
+
+  background-color: var(--primary-color);
+`;
+
+export const Title = styled.h1`
+  ${textTitle}
+  color: #fcfcfd;
+`;
+
+export const AddRestaurantButton = styled.button`
+  height: 40px;
+
+  border: none;
+  border-radius: 8px;
+  background: transparent;
+
+  font-size: 24px;
+  cursor: pointer;
+
+  img {
+    display: block;
+    width: 40px;
+    height: 40px;
+    object-fit: contain;
+  }
+`;
