+++ src/components/index.ts
@@ -1,8 +1,8 @@
 import Button from './Button';
-import CategoryImage from './CategoryImage';
-import Header from './Header';
-import RestaurantDetail from './RestaurantDetail';
-import RestaurantItem from './RestaurantItem';
-import RestaurantList from './RestaurantList';
+import CategoryImage from './CategoryImage/CategoryImage';
+import Header from './Header/Header';
+import RestaurantDetail from './RestaurantDetail/RestaurantDetail';
+import RestaurantItem from './RestaurantItem/RestaurantItem';
+import RestaurantList from './RestaurantList/RestaurantList';
 
 export { Button, CategoryImage, Header, RestaurantDetail, RestaurantList, RestaurantItem };
